+++ packages/ethereum/sdk/src/config/polygon-amoy.ts
@@ -3,17 +3,15 @@ import { id32 } from "../common/id"
 import type { EthereumConfig } from "./type"
 import { FEE_CONFIG_URL } from "./common"
 
-export const testnetEthereumConfig: EthereumConfig = {
-	basePath: "https://testnet-ethereum-api.rarible.org",
-	chainId: 11155111,
+export const polygonAmoyConfig: EthereumConfig = {
+	basePath: "https://testnet-polygon-api.rarible.org",
+	chainId: 80002,
 	environment: "testnet",
 	exchange: {
 		v1: ZERO_ADDRESS,
 		v2: toAddress("0x3e52D660b69d1bDacb6C513cE085D924F5Cb9c77"),
 		openseaV1: ZERO_ADDRESS,
-		wrapper: toAddress("0x1554aDA53194B961016931A2E86C80D09a816209"),
-		looksrare: ZERO_ADDRESS,
-		looksrareV2: toAddress("0x34098cc15a8a48Da9d3f31CC0F63F01f9aa3D9F3"),
+		wrapper: toAddress("0x2FCE8435F0455eDc702199741411dbcD1B7606cA"),
 		x2y2: ZERO_ADDRESS,
 	},
 	transferProxies: {
@@ -28,21 +26,19 @@ export const testnetEthereumConfig: EthereumConfig = {
 	openSea: {
 		metadata: id32("RARIBLE"),
 		proxyRegistry: ZERO_ADDRESS,
-		merkleValidator: ZERO_ADDRESS,
 	},
 	factories: {
-		erc721: toAddress("0xB020bA7fcF43DCc59eF0103624BD6FADE66d105E"),
-		erc1155: toAddress("0x166F6180170f438Ddc38050a2B708d38c0890956"),
+		erc721: toAddress("0x18a2553ef1aaE12d9cd158821319e26A62feE90E"),
+		erc1155: toAddress("0xc9eB416CDb5cc2aFC09bb75393AEc6dBA4E5C84a"),
 	},
 	cryptoPunks: {
 		marketContract: ZERO_ADDRESS,
 		wrapperContract: ZERO_ADDRESS,
 	},
 	sudoswap: {
-		pairFactory: toAddress("0xd96B8cd321176D95C77B2Ba6bfC007659c6CdceB"),
-		pairRouter: toAddress("0x72d0Ee6B28553b048442a9c8DAD6eA33806e9357"),
+		pairFactory: ZERO_ADDRESS,
+		pairRouter: ZERO_ADDRESS,
 	},
-	weth: toAddress("0xfff9976782d46cc05630d1f6ebab18b2324d6b14"),
+	weth: toAddress("0x0ae690aad8663aab12a671a6a0d74242332de85f"),
 	auction: ZERO_ADDRESS,
-	looksrareOrderValidatorV2: toAddress("0x0bc129E4c1f8D7b5583eAbAeb1F7468935B6ec0C"),
 }