+++ packages/ethereum/sdk/src/order/fill-order/looksrare.test.ts
@@ -18,7 +18,7 @@ import { createE2eTestProvider } from "../../common/test/create-test-providers"
 import { makeRaribleSellOrder } from "./looksrare-utils/create-order"
 
 describe.skip("looksrare fill", () => {
-	const goerli = getE2EConfigByNetwork("goerli")
+	const goerli = getE2EConfigByNetwork("sepolia")
 	const { provider: providerBuyer } = createE2eProvider(DEV_PK_1, goerli)
 	const { provider: providerSeller } = createE2eProvider(DEV_PK_2, goerli)
 	const { wallet: feeWallet } = createE2eProvider(undefined, goerli)