+++ packages/transaction/src/ethereum/index.ts
@@ -41,7 +41,7 @@ Blockchain, TransactionResult
 			case "polygon":
 				return `https://polygonscan.com/tx/${this.hash()}`
 			case "testnet":
-				return `https://goerli.etherscan.io/tx/${this.hash()}`
+				return `https://sepolia.etherscan.io/tx/${this.hash()}`
 			case "mantle":
 				return `https://explorer.mantle.xyz/tx/${this.hash()}`
 			case "testnet-mantle":
@@ -74,6 +74,8 @@ Blockchain, TransactionResult
 				return `http://ethereum-explorer.dev.rarible.int/transactions/${this.hash()}`
 			case "dev-polygon":
 				return `http://polygon-explorer.dev.rarible.int/transactions/${this.hash()}`
+			case "amoy-polygon":
+				return `https://amoy.polygonscan.com/tx/${this.hash()}`
 			case "zkatana":
 				return `https://zkatana.blockscout.com/tx/${this.hash()}`
 			case "astar-zkevm":